Moved files to a new folder to maintain better order
