+++ components/features/programDirector/partialTemplate/domain/entities/PartialTemplate.ts
